Pulled all iteration game stats test and methods into GameStats class
